Fix multiple initialization
If multiple libraries get combined that initialize the master sink then
it's possible for it to get initialized multiple times during static
initialization (and then one of the "wins" and becomes the global
variable).  This change makes sure that if that happen, they end up
referring to the same instance.