Fix multiple path strip roots, and prepend new roots
PARENT_SCOPE wasn't working to preserve the variable, so store it in an
internal CACHE variable instead, which seems to work.

Prepending to the list allows a caller to manipulate the paths a bit
(e.g. putting subdirectories later than parent directories, so that the
subdirectories get stripped rather than just the parent directories).